+++ src/modules/ttlmonitor/CMakeLists.txt
@@ -30,18 +30,18 @@ create_module(ttlmonitor)
 
 
 #
-#set(TTLMonitorBaseTest_INC
-#  ${ttlmonitor_common_inc}
-#  ${ESS_SOURCE_DIR}/common/testutils/SaveBuffer.h
-#  ${ESS_SOURCE_DIR}/common/testutils/TestUDPServer.h
-#)
-#set(TTLMonitorBaseTest_SRC
-#  TTLMonitorBaseTest.cpp
-#  ${ESS_SOURCE_DIR}/common/testutils/SaveBuffer.cpp
-#  ${freia_common_src}
-#  ${ESS_SOURCE_DIR}/common/testutils/TestUDPServer.cpp
-#)
-#create_test_executable(TTLMonitorBaseTest)
+set(TTLMonitorBaseTest_INC
+  ${ttlmonitor_common_inc}
+  ${ESS_SOURCE_DIR}/common/testutils/SaveBuffer.h
+  ${ESS_SOURCE_DIR}/common/testutils/TestUDPServer.h
+)
+set(TTLMonitorBaseTest_SRC
+  test/TTLMonitorBaseTest.cpp
+  ${ESS_SOURCE_DIR}/common/testutils/SaveBuffer.cpp
+  ${ttlmonitor_common_src}
+  ${ESS_SOURCE_DIR}/common/testutils/TestUDPServer.cpp
+)
+create_test_executable(TTLMonitorBaseTest)
 
 
 #
